backend/requirements.txt
fastapi
uvicorn[standard]
httpx
sse-starlette
python-dotenv
pydantic>=2.0
